Completed Camera functionality and highlighting tool
essentially camera is now zoomable from the mouse scroll wheel and moveable via WASD/arrow keys so players can see and move around the map.
Also the specific tile of the tilemap that the cursor is on now is highlighted by a animated, transparent shader. can change in future to a animated red/black outline of the square later if that suits better.