Added UI counters for credits and round + next wave button
added simple UI counters for credits and the round (it just uses basic text no fancy pixel art UI yet). i set it up so there is a button which increases the round, and each round the credits increases (i set it up to add 100 but we can change the amount). the round number and credits is updated to the player_saving script.

my unity keeps crashing but the scene was saved under Assets > Recovery > 0 (5). i havent moved it in replacement of the original Map scene because i dont want to stuff anything up